requirements.txt
flask
pandas
requests
rich
python-dotenv